requirements.txt
python-telegram-bot~=12.8
PyYAML~=5.3.1
jellyfish~=0.8.2
